Chore: Sync files

+++ migrations/2023-08-20-032217_create_organization/up.sql
@@ -3,12 +3,12 @@ CREATE TABLE organizations (
     id TEXT PRIMARY KEY NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     -- updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    clock DOUBLE NOT NULL,
-    cpi_instruction_r DOUBLE NOT NULL,
-    cpi_instruction_i DOUBLE NOT NULL,
-    cpi_instruction_l DOUBLE NOT NULL,
-    cpi_instruction_s DOUBLE NOT NULL,
-    cpi_instruction_b DOUBLE NOT NULL,
-    cpi_instruction_u DOUBLE NOT NULL,
-    cpi_instruction_j DOUBLE NOT NULL
+    clock FLOAT NOT NULL,
+    cpi_instruction_r FLOAT NOT NULL,
+    cpi_instruction_i FLOAT NOT NULL,
+    cpi_instruction_l FLOAT NOT NULL,
+    cpi_instruction_s FLOAT NOT NULL,
+    cpi_instruction_b FLOAT NOT NULL,
+    cpi_instruction_u FLOAT NOT NULL,
+    cpi_instruction_j FLOAT NOT NULL
 );
